requirements.txt
streamlit>=1.32.0
plotly>=5.19.0
nltk>=3.8.1
fpdf2>=2.7.9
numpy>=1.26.0
